Removing fixed color and setting point / rectangle classes data to int

+++ src/main/Point.java
@@ -2,15 +2,15 @@ package main;
 
 public class Point {
 
-	public double x;
-	public double y;
+	public int x;
+	public int y;
 	
 	public Point() {
 		this.x = 0;
 		this.y = 0;
 	}
 	
-	public Point(double x, double y) {
+	public Point(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
